Bug 1290830 - [Part3] Fix timestamp type from double to int64_t for ExpirationChange in PGMPDecryptor.ipdl. r?cpearce
MozReview-Commit-ID: 4dznu2Md1r3

+++ dom/media/gmp/GMPDecryptorParent.h
@@ -89,7 +89,7 @@ private:
                           nsTArray<uint8_t>&& aMessage) override;
 
   bool RecvExpirationChange(const nsCString& aSessionId,
-                            const double& aExpiryTime) override;
+                            const int64_t& aExpiryTime) override;
 
   bool RecvSessionClosed(const nsCString& aSessionId) override;
